Fixed Bullet System + Enemy Group Brainstorm
Changed the previous projectile system to that of an array, allowing it to be more compatible with future additions, and reimplemented said system into the working items. Also added a filter to catch stray bullets. Added scratch comment code to plan rebuild of enemy groupings in order to improve code compatibility.